Split agriculture.rs (709 lines) into subdirectory modules
Closes #1389

Split into focused sub-modules:
- types.rs: constants, FrostEvent, AgricultureState
- helpers.rs: pure calculation functions (frost risk, temperature suitability, rainfall adequacy)
- systems.rs: update_agriculture system and AgriculturePlugin
- tests.rs: all unit tests

All public items re-exported from mod.rs so external callers are unaffected.
