+++ crates/simulation/src/agriculture/mod.rs
@@ -1,0 +1,11 @@
+mod helpers;
+mod systems;
+mod tests;
+pub mod types;
+
+pub use helpers::{
+    calculate_crop_yield, calculate_frost_risk, is_growing_season, rainfall_adequacy,
+    temperature_suitability,
+};
+pub use systems::{update_agriculture, AgriculturePlugin};
+pub use types::{AgricultureState, FrostEvent};
